SBERDOMA-1012 add accessSchema
