+++ apps/condo/domains/meter/access/Meter.js
@@ -1,111 +1,24 @@
 // @ts-nocheck
+const { canReadMeterEntity } = require('@condo/domains/meter/utils/accessSchema')
 /**
  * Generated by `createschema meter.Meter 'number:Text; billingAccountMeter?:Relationship:BillingAccountMeter:SET_NULL; account?:Relationship:BillingAccount:SET_NULL; property:Relationship:Property:CASCADE; unitName:Text; place?:Text; resource:Relationship:MeterResource:CASCADE;'`
  */
 const { Meter } = require('../utils/serverSchema')
-const { Property } = require('@condo/domains/property/utils/serverSchema')
-const { queryOrganizationEmployeeFromRelatedOrganizationFor } = require('@condo/domains/organization/utils/accessSchema')
-const { queryOrganizationEmployeeFor } = require('@condo/domains/organization/utils/accessSchema')
-const { RESIDENT } = require('@condo/domains/user/constants/common')
-const get = require('lodash/get')
-const { throwAuthenticationError } = require('@condo/domains/common/utils/apolloErrorFormatter')
-const { checkRelatedOrganizationPermission } = require('../../organization/utils/accessSchema')
-const { getById } = require('@core/keystone/schema')
-const { checkOrganizationPermission } = require('@condo/domains/organization/utils/accessSchema')
+const { canManageMeterEntity } = require('@condo/domains/meter/utils/accessSchema')
 
 async function canReadMeters ({ authentication: { item: user } }) {
-    if (!user) return throwAuthenticationError()
-    if (user.isAdmin || user.isSupport) {
-        return {}
-    }
-    if (user.type === RESIDENT) {
-        return {
-            createdBy: { id: user.id },
-        }
-    }
-    const userId = user.id
-    return {
-        organization: {
-            OR: [
-                queryOrganizationEmployeeFor(userId),
-                queryOrganizationEmployeeFromRelatedOrganizationFor(userId),
-            ],
-        },
-    }
+    return await canReadMeterEntity({ user })
 }
 
 async function canManageMeters ({ authentication: { item: user }, originalInput, operation, itemId, context }) {
-    if (!user) return throwAuthenticationError()
-    if (user.isAdmin || user.isSupport) return true
-    if (operation === 'create') {
-        const organizationIdFromMeter = get(originalInput, ['organization', 'connect', 'id'])
-        if (!organizationIdFromMeter) {
-            return false
-        }
-
-        const propertyId = get(originalInput, ['property', 'connect', 'id'])
-
-        const [property] = await Property.getAll(context, { id: propertyId })
-        if (!property) {
-            return false
-        }
-        if (user.type === RESIDENT) {
-            return true
-        }
-        const canManageRelatedOrganizationTickets = await checkRelatedOrganizationPermission(context, user.id, organizationIdFromMeter, 'canManageMeters')
-        if (canManageRelatedOrganizationTickets) {
-            return true
-        }
-        const organizationIdFromProperty = get(property, ['organization', 'id'])
-        const canManageMeters = await checkOrganizationPermission(context, user.id, organizationIdFromMeter, 'canManageMeters')
-        if (!canManageMeters) {
-            return false
-        }
-
-        return organizationIdFromMeter === organizationIdFromProperty
-
-    } else if (operation === 'update') {
-        if (!itemId) {
-            return false
-        }
-        const meter = await Meter.getAll(context, { id: itemId })
-        if (!meter) {
-            return false
-        }
-        if (meter.createdBy === user.id && user.type === RESIDENT) {
-            return true
-        }
-
-        const { organization: organizationIdFromMeter } = meter
-
-        const canManageRelatedOrganizationTickets = await checkRelatedOrganizationPermission(context, user.id, organizationIdFromMeter, 'canManageMeters')
-        if (canManageRelatedOrganizationTickets) {
-            return true
-        }
-        const canManageMeters = await checkOrganizationPermission(context, user.id, organizationIdFromMeter, 'canManageMeters')
-        if (!canManageMeters) {
-            return false
-        }
-
-        const propertyId = get(originalInput, ['property', 'connect', 'id'])
-        if (propertyId) {
-            const [property] = await Property.getAll(context, { id: propertyId })
-            if (!property) {
-                return false
-            }
-
-            const organizationIdFromProperty = get(property, ['organization', 'id'])
-            const isSameOrganization = organizationIdFromMeter === organizationIdFromProperty
-
-            if (!isSameOrganization) {
-                return false
-            }
-        }
-
-        return true
-    }
-
-    return false
+    return await canManageMeterEntity({
+        schema: Meter,
+        user,
+        itemId,
+        operation,
+        originalInput,
+        context,
+    })
 }
 
 /*
